infinite loop bug -- close branch
